Do not switch groups on the session when linking annotations
There is no longer a reason to switch groups on the session when linking
annotations to the current object.  This causes all manner of
concurrency issues, especially when there are open stateful services, as
`setGroupForSession()` actively closes all open stateful services (it is
required to do so).

The current implementation of `_linkObject()` handles the assignment of
group perfectly via `SERVICE_OPTS`.